backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
Pillow>=10.0
python-dotenv>=1.0
